codex: colocate user instruction integration test
