+++ codex-rs/core/tests/suite/mod.rs
@@ -120,7 +120,6 @@ mod unified_exec;
 #[cfg(unix)]
 mod unified_exec_zsh_fork_approvals;
 mod unstable_features_warning;
-mod user_instructions;
 mod user_notification;
 mod user_shell_cmd;
 mod view_image;
